+++ tests/app/models.py
@@ -86,6 +86,9 @@ class Region(models.Model):
     name = models.CharField(max_length=200)
     mayor = models.OneToOneField(Person, null=True, on_delete=models.CASCADE)
 
+    class Meta:
+        ordering = ["name"]
+
     def __str__(self):
         return self.name